Prop 05
"equal multiple" is likely meant just about every point at which "equal" is used (same too whereat "multiple" is used) however the end of the second sentence reads that f.d will equal a.b and that equality is mirrored in the numerical example of the diagram (though not in its literal representation as a line; the number was provided by Campanus afaict and the line was just what the printers could fit inside the margin, I'm assuming). a subsequent printing of this text from 1558 changes that line to read that c.d will equal a.b, which would be correct insofar as c.d and a.b are "equal multiples"; the 1558-printing forgoes the numbers (probably a good call) but its diagram depicts the line f.d as being visually thereabout equal to the line a.b, though this may just be a funny(?) coincidence.

"quota," so far as I can tell, is newly being used as a noun here, namely, a part of a pro/portion.